+++ wampum.php
@@ -165,7 +165,7 @@ final class Wampum_Setup {
 
 		// Plugin version.
 		if ( ! defined( 'WAMPUM_VERSION' ) ) {
-			define( 'WAMPUM_VERSION', '1.4.7.1' );
+			define( 'WAMPUM_VERSION', '1.4.8' );
 		}
 
 		// Plugin Folder Path.
@@ -292,7 +292,7 @@ final class Wampum_Setup {
 	 		'parent_slug'  => 'themes.php',             // Parent menu slug.
 	 		'capability'   => 'edit_theme_options',     // Capability needed to view plugin install page, should be a capability associated with the parent menu used.
 	 		'has_notices'  => true,                     // Show admin notices or not.
-	 		'dismissable'  => false,                    // If false, a user cannot dismiss the nag message.
+	 		'dismissable'  => true,                     // If false, a user cannot dismiss the nag message.
 	 		'dismiss_msg'  => '',                       // If 'dismissable' is false, this message will be output at top of nag.
 	 		'is_automatic' => false,                    // Automatically activate plugins after installation or not.
 	 		'message'      => '',                       // Message to output right before the plugins table.